requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-huggingface
gradio
sentence-transformers
torch
transformers
faiss-cpu
python-dotenv
requests
pypdf
beautifulsoup4
lxml